requirements.txt
requests
pyyaml
pytest
